# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bookstore API",
        "description": "API documentation for Bookstore project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local server"
        }
    ],
    "paths": {
        "/api/authors/top": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Get top authors",
                "description": "Retrieve a ranked list of top authors based on popularity, average rating, or trending score.",
                "operationId": "4d44841afe83b2bc38dbfc31a431a688",
                "parameters": [
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort mode: popularity, average, or trending",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "popularity",
                                "average",
                                "trending"
                            ]
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of top authors to return (default: 10)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Top authors retrieved successfully"
                    },
                    "400": {
                        "description": "Invalid query parameters"
                    }
                }
            }
        },
...
    },
...
}